Remove extra check for host
I'm using MySQL with a ClearDB package on Heroku. The ClearDB credentials come as an URL. The MySQL package is able to handle this url directly e.g. this works: 
````
var connection = mysql.createConnection(process.env["CLEARDB_DATABASE_URL"]);
````
However, this package (erroneously) checks if the host property is set. This prevents me from connecting to MySQL with an URL.